불필요한 파일 제거 및 코드 정리
