--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "seoul42-e84e5"
-  }
-}
