+++ .gitignore
@@ -35,6 +35,3 @@ jspm_packages
 
 # Optional REPL history
 .node_repl_history
-
-#service, api key
-routes/utils/config/*.json
