--- firebase.json
@@ -1,20 +1,0 @@
-{
-  "functions": {
-    "predeploy": [],
-    "source": "functions" 
-  },
-  "hosting": {
-    "public": "public",
-    "ignore": [
-      "firebase.json",
-      "**/.*",
-      "**/node_modules/**"
-    ],
-    "rewrites":[
-    	{
-        "source": "**",
-        "function": "apps"
-      }
-    ]
-  }
-}
