+++ index.html
@@ -1,0 +1,87 @@
+<h1 id="imgwidth85altkakaotalk_20190212_204520833srchttpsuserimagesgithubusercontentcom22374750528346261efe2080312611e992c366ac08c1f2c9png">서울 살이 <img width="85" alt="kakaotalk_20190212_204520833" src="https://user-images.githubusercontent.com/22374750/52834626-1efe2080-3126-11e9-92c3-66ac08c1f2c9.png"></h1>
+
+<p><img src="https://img.shields.io/badge/service-android-green.svg" alt="Seoul42 Service" />
+<img src="https://img.shields.io/badge/pattern-MVVM-blue.svg" alt="Seoul42 Pattern" />
+<a href="https://insert-koin.io"><img src="https://img.shields.io/badge/DI-koin-blue.svg" alt="Seoul42 DI" /></a>
+<a href="https://app.zeplin.io/project/5c4db2597a8bebbfe8be9d39/dashboard"><img src="https://img.shields.io/badge/UI-zeplin-blue.svg" alt="Seoul42_UI" /></a></p>
+
+<p><a href="https://github.com/amitshekhariitbhu/RxJava2-Android-Samples"><img src="https://img.shields.io/badge/library-RxJava2-blue.svg" alt="Seoul42_UI" /></a>
+<a href="https://developer.android.com/topic/libraries/architecture/room.html"><img src="https://img.shields.io/badge/library-Room-blue.svg" alt="Seoul42_UI" /></a>
+<a href="https://developer.android.com/topic/libraries/architecture/livedata.html"><img src="https://img.shields.io/badge/library-LiveData-blue.svg" alt="Seoul42_UI" /></a></p>
+
+<p>booscamp3_C팀의 서울살이 repository입니다. 해당 프로젝트는 MVVM기반의 Android프로젝트입니다. 
+의존성 주입을 위해서 koin이 사용되었고, RxJava, Room, LiveData, MotionLayout등 꾸준한 학습에 의해서 진행되고 있습니다.</p>
+
+<hr/>
+
+<h4 id="">팀원 소개 : 문병학, 유지원, 최준영(링크 추가 예정)</h4>
+
+<p>서울살이는 서울살이를 시작하는 사람들에게 주택에 관한 정보를 제공하는 어플리케이션입니다.</p>
+
+<p>현재 아파트에 대한 정보를 제공하는 다양한 어플리케이션이 존재하지만, 상대적으로 주택에 관련된 정보를 제공하는 어플리케이션은 부족합니다.
+그렇기 때문에 이러한 부족한 정보를 서울살이라는 어플리케이션을 통해서 주택에 관련된 전/월세, 주변 지역 정보, 주택 정보등의 정보를 제공하려고 합니다.</p>
+
+<hr/>
+
+<h4 id="uiseoul42_uihttpsimgshieldsiobadgeuizeplinbluesvghttpsappzeplinioproject5c4db2597a8bebbfe8be9d39dashboard">프로토타입 UI: <a href="https://app.zeplin.io/project/5c4db2597a8bebbfe8be9d39/dashboard"><img src="https://img.shields.io/badge/UI-zeplin-blue.svg" alt="Seoul42_UI" /></a></h4>
+
+<p><img src="https://user-images.githubusercontent.com/22374750/52126227-17dd0a00-2672-11e9-9678-2cf1e2aeb6fc.png" alt="image" /></p>
+
+<p>해당 프로젝트는 협업툴을 Zeplin을 사용했습니다. 구체적인 내용은 해당 뱃지를 누르면 확인할 수 있습니다.<br>
+저희는 Zeplin을 통해서 Style 가이드를 작성했고, 각종 xml의 가이드 기준을 정했습니다.</p>
+
+<h4 id="-1">패키지 소개:</h4>
+
+<ol>
+<li><strong>api</strong>: Retrofit을 사용하기 위한 interface를 모아두는 패키지입니다.</li>
+
+<li><strong>di</strong>: Koin을 사용하여 의존성 주입을 했고, 해당 패키지에 di 관련 모듈을 만들었습니다.</li>
+
+<li><strong>firebase</strong>: Firebase에 관련한 유틸리티성 클래스를 모아두는 패키지입니다.</li>
+
+<li><strong>model</strong>: 주소, 북마크, 집정보, 가격등에 관련한 데이터 클래스를 모아두는 패키지입니다.</li>
+
+<li><strong>repository</strong>: ViewModel과 Model사이에 repository패턴을 사용해서 각종 값을 받아오는 클래스를 모아둔 패키지입니다.</li>
+
+<li><strong>ui</strong>: View에 해당하는 클래스들을 모아두는 패키지입니다.</li>
+
+<li><strong>util</strong>: 프로젝트내에서 각종 유틸리티성 클래스를 모아두는 패키지입니다.</li>
+</ol>
+
+<h3 id="httpsdrivegooglecomfiled1ui4lvmc81kcaki4uvtxirsg0szlebqd2viewuspsharing"><a href="https://drive.google.com/file/d/1ui4lvMc81kCAki4UVtxirsg0szLEbqD2/view?usp=sharing">기획안</a></h3>
+
+<h3 id="httpsdocsgooglecomspreadsheetsd1nqlae8oneo42rk9pr0tzfxilmcsfunrmooc4p7cflyyedituspsharing"><a href="https://docs.google.com/spreadsheets/d/1nQlae8ONeO42Rk9Pr0tZFxilmCsfuNRmOOc4p7cFlYY/edit?usp=sharing">일정관리</a></h3>
+
+<h3 id="uihttpsdrivegooglecomfiled13bltmr3iynhjudikymlrvolux6oqniuviewuspsharing"><a href="https://drive.google.com/file/d/13BLtMr3i-YnhjuDIkYmLRVolUX6OQnIu/view?usp=sharing">UI기획</a></h3>
+
+<h3 id="httpsdocsgooglecomspreadsheetsd1y4xpb8lsp5qq53e1npewszmmyud5io1h1sqxxzzomy4edituspsharing"><a href="https://docs.google.com/spreadsheets/d/1Y4Xpb8lSP5qQ53e1NPewsZMmYud5io1H1SQxxZZOmY4/edit?usp=sharing">기능명세서</a></h3>
+
+<h3 id="httpsgithubcomseoul42seoul42server"><a href="https://github.com/seoul42/seoul42-server">서버</a></h3>
+
+<hr/>
+
+<h2 id="1">1주차 산출물</h2>
+
+<p><a href="https://github.com/boostcampth/boostcamp3_C/tree/dev/docs">DOCS</a> 문서 작업 (기획서, 기능정의서, 프로젝트 일정, Api명세서)</p>
+
+<p>Zeplin<a href="https://app.zeplin.io/project/5c4db2597a8bebbfe8be9d39/dashboard"><img src="https://img.shields.io/badge/UI-zeplin-blue.svg" alt="Seoul42_UI" /></a>
+<img src="https://user-images.githubusercontent.com/22374750/52126227-17dd0a00-2672-11e9-9678-2cf1e2aeb6fc.png" alt="image" /></p>
+
+<h2 id="2">2주차 산출물</h2>
+
+<p>Seoul42 API <a href="https://app.zeplin.io/project/5c4db2597a8bebbfe8be9d39/dashboard"><img src="https://img.shields.io/badge/API-SwaggerHub-blue.svg" alt="Seoul42_API" /></a>
+<img src="https://user-images.githubusercontent.com/22374750/52177071-3f151200-27ff-11e9-8b58-ebc916b21c75.png" alt="image" /></p>
+
+<h3 id="2-1">2주차 회고록</h3>
+
+<p><img src="https://user-images.githubusercontent.com/22374750/52834631-26bdc500-3126-11e9-9b4e-1c2eb4ff5895.jpg" alt="kakaotalk_20190214_181143100" />
+<img src="https://user-images.githubusercontent.com/22374750/52834632-26bdc500-3126-11e9-8f68-95c67e153998.jpg" alt="kakaotalk_20190214_181143373" />
+준영님 파일 재업로드 필요</p>
+
+<h2 id="3">3주차 산출물</h2>
+
+<h3 id="3-1">3주차 회고록</h3>
+
+<p><img src="https://user-images.githubusercontent.com/22374750/52834634-27565b80-3126-11e9-98fa-45f600706af6.jpg" alt="kakaotalk_20190214_181144018" />
+<img src="https://user-images.githubusercontent.com/22374750/52834636-27565b80-3126-11e9-8b02-ccb4046a729f.jpg" alt="kakaotalk_20190214_181144329" />
+<img src="https://user-images.githubusercontent.com/22374750/52834637-27565b80-3126-11e9-8a4c-f851f327d334.jpg" alt="kakaotalk_20190214_181144629" /></p>
