--- nodejs-server-server-generated/.swagger-codegen/VERSION
@@ -1,1 +1,0 @@
-2.4.0
