--- nodejs-server-server-generated/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "seoul42-server-api",
-  "version": "1.0.0",
-  "description": "서울살이 API에 대한 RestAPI 명세서입니다.",
-  "main": "index.js",
-  "scripts": {
-	"prestart": "npm install",
-	"start": "node index.js"
-  },
-  "keywords": [
-    "swagger"
-  ],
-  "license": "Unlicense",
-  "private": true,
-  "dependencies": {
-    "connect": "^3.2.0",
-    "js-yaml": "^3.3.0",
-    "swagger-tools": "0.10.1"
-  }
-}
