+++ package.json
@@ -4,19 +4,16 @@
   "main": "main.js",
   "scripts": {
     "build": "echo 'hello build'",
-    "start": "node server.js"
+    "start": "node index.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/proj4": "^2.5.0",
     "async": "^2.6.1",
     "body-parser": "^1.18.3",
     "connect": "^3.6.6",
     "cors": "^2.8.5",
     "express": "^4.16.4",
-    "firebase-admin": "^7.0.0",
-    "firebase-functions": "^2.2.0",
     "js-yaml": "^3.12.1",
     "natives": "^1.1.6",
     "ol": "^5.3.0",
@@ -26,7 +23,7 @@
     "swagger-tools": "^0.10.4",
     "urlencode": "^1.1.0"
   },
+  "engines": { "node": "^10.14.1", "npm": "^6.4.1" },
   "devDependencies": {},
-  "description": "",
-  "private": true
+  "description": ""
 }
