--- routes/utils/config/.gitignore
@@ -1,1 +1,0 @@
-*.json
